// src/app/core/api/model/warehouse.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Tenant } from './tenant';
import { Branch } from './branch';


export interface Warehouse { 
    id?: number;
    name?: string;
    description?: string;
    branch?: Branch;
    tenant?: Tenant;
    'default'?: boolean;
}