// src/app/core/api/model/inventory.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Warehouse } from './warehouse';
import { Product } from './product';


export interface Inventory { 
    id?: number;
    warehouse?: Warehouse;
    product?: Product;
    quantity?: number;
}